added vanished changes
