+++ src/domain/hanke/edit/Form1.tsx
@@ -1,14 +1,13 @@
-import React, { Dispatch, SetStateAction } from 'react';
+import React from 'react';
 import { Button } from 'hds-react';
 import { useTranslation } from 'react-i18next';
 
-interface IProps {
-  changeWizardView: Dispatch<SetStateAction<number>>;
-}
+import PropTypes from './PropTypes';
 
-const Form1: React.FC<IProps> = (props) => {
-  const { t } = useTranslation();
+const Form1: React.FC<PropTypes> = (props) => {
   const { changeWizardView } = props;
+
+  const { t } = useTranslation();
   return (
     <div className="form1">
       <h2>{t('hankeForm:hankkeenAlueForm:header')}</h2>
